+++ typescript/eslint-config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/eslint-config
 
+## 29.1.0
+
 ## 29.0.1
 
 ## 29.0.0